Rename test directory to match new lib name
