+++ test/over_react_test.html
@@ -6,7 +6,7 @@
     <script src="packages/react/react_with_addons.js"></script>
     <script src="packages/react/react_dom.js"></script>
 
-    <link rel="x-dart-test"  href="test_util_test.dart">
+    <link rel="x-dart-test"  href="over_react_test.dart">
     <script src="packages/test/dart.js"></script>
 </head>
 <body></body>
